fixed bugs and warnings
git-svn-id: http://svn.seqan.de/seqan/trunk@7970 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/benchmarks/read_mappers/apps/compare_sam_wit.h
@@ -282,7 +282,7 @@ int bestScoreForAligned(TFragmentStore & fragments,
 
 // Return maximum error count for maximum error rate
 inline int maxErrorRateToMaxErrors(int maxErrorRate, size_t len) {
-    return floor(maxErrorRate / 100.0 * len);
+    return (int)floor(maxErrorRate / 100.0 * len);
 }
